+++ CHANGELOG.md
@@ -6,6 +6,17 @@ List all changes in various categories:
 * Bug fixes
 * Known issues
 
+## 2.0.0 (2018-09-16) ##
+
+### Release notes ###
+
+- OAuth2 token is now provided by default in Authorization header.
+
+### Bug fixes ###
+
+- Oauth2 authentication was not working with Python 2.7
+- Update requests to 2.19.1 (latest version)
+
 ## 1.0.2 (2018-01-19) ##
 
 ### Release notes ###